Updates homepage to new styles
